Most recent CL version 18/02
I have uploaded the file through "file upload" instead of pushing it into the repository. The reason for that is that I wanted to have the most recent version that has been tested on a University Lab machine. This version successfully runs all three commands of the Start Menu and executes what it is supposed to.
~Ivaylo Dimanov